Auto-filling complete!
This is a big update! I've implemented algorithms that project known frequency values at various temperatures through the temperatures where those frequencies might be missing for whatever reason. The method for doing this is documented in utilities. I've also added a GUI for interacting with this in the mistake selector window.


Former-commit-id: 079531302557e66a1a732815f9aaaed3bb312854 [formerly 0c56ea77abc10a7ce12c98ec1c27e5dd4270d744] [formerly 9e98a9ec6b36e3e5d2246678a27ffd300ecbac8f [formerly b4bcce5c6e1f066d1b5327d8a0d447cde1873cd5]]
Former-commit-id: 7caf7e53db8f39336915bc19e52ed2c08760723c [formerly d279da1f1415336a36e658004c6218fdccf74bcb]
Former-commit-id: 6fa07e867977c7c62bc2d894da4521c4f498d67c